{"schema":"decodex/commit/1","summary":"Split consolidation service modules","authority":"manual"}
